azure_functions/inference/requirements.txt
azure-functions
pandas
numpy
scikit-learn
joblib
azure-storage-blob
mlflow
